refactor: split paths code into separate file
