+++ packages/aql/src/compiler/types.ts
@@ -1,0 +1,9 @@
+export type PathInfo = {
+    tableName: string,
+    tableId: string,
+    joinField: string,
+    joinTable: string,
+    noMapping?: boolean
+}
+
+export type CompilerState = any;
